requirements.txt
discord.py
numpy
scikit-learn
matplotlib
pandas
pymongo[srv]
requests